+++ lib/searchkick/query.rb
@@ -16,9 +16,9 @@ module Searchkick
 
     def initialize(klass, term = "*", **options)
       unknown_keywords = options.keys - [:aggs, :body, :body_options, :boost,
-        :boost_by, :boost_by_distance, :boost_where, :conversions, :debug, :emoji, :exclude, :execute, :explain,
+        :boost_by, :boost_by_distance, :boost_where, :conversions, :conversions_term, :debug, :emoji, :exclude, :execute, :explain,
         :fields, :highlight, :includes, :index_name, :indices_boost, :limit, :load,
-        :match, :misspellings, :offset, :operator, :order, :padding, :page, :per_page, :profile,
+        :match, :misspellings, :model_includes, :offset, :operator, :order, :padding, :page, :per_page, :profile,
         :request_params, :routing, :select, :similar, :smart_aggs, :suggest, :track, :type, :where]
       raise ArgumentError, "unknown keywords: #{unknown_keywords.join(", ")}" if unknown_keywords.any?
 
@@ -79,7 +79,7 @@ module Searchkick
       @execute ||= begin
         begin
           response = execute_search
-          if @misspellings_below && response["hits"]["total"] < @misspellings_below
+          if retry_misspellings?(response)
             prepare
             response = execute_search
           end
@@ -108,6 +108,7 @@ module Searchkick
         padding: @padding,
         load: @load,
         includes: options[:includes],
+        model_includes: options[:model_includes],
         json: !@json.nil?,
         match_suffix: @match_suffix,
         highlighted_fields: @highlighted_fields || [],
@@ -157,6 +158,10 @@ module Searchkick
 
       # set execute for multi search
       @execute = Searchkick::Results.new(searchkick_klass, response, opts)
+    end
+
+    def retry_misspellings?(response)
+      @misspellings_below && response["hits"]["total"] < @misspellings_below
     end
 
     private
@@ -218,6 +223,11 @@ module Searchkick
 
       @json = options[:body]
       if @json
+        ignored_options = options.keys & [:aggs, :boost,
+          :boost_by, :boost_by_distance, :boost_where, :conversions, :conversions_term, :exclude, :explain,
+          :fields, :highlight, :indices_boost, :limit, :match, :misspellings, :offset, :operator, :order,
+          :padding, :page, :per_page, :select, :smart_aggs, :suggest, :where]
+        warn "The body option replaces the entire body, so the following options are ignored: #{ignored_options.join(", ")}" if ignored_options.any?
         payload = @json
       else
         if options[:similar]
@@ -381,7 +391,7 @@ module Searchkick
                       boost_mode: "replace",
                       query: {
                         match: {
-                          "#{conversions_field}.query" => term
+                          "#{conversions_field}.query" => options[:conversions_term] || term
                         }
                       }
                     }.merge(script_score)
@@ -447,7 +457,7 @@ module Searchkick
         set_aggregations(payload) if options[:aggs]
 
         # suggestions
-        set_suggestions(payload) if options[:suggest]
+        set_suggestions(payload, options[:suggest]) if options[:suggest]
 
         # highlight
         set_highlights(payload, fields) if options[:highlight]
@@ -468,14 +478,15 @@ module Searchkick
         elsif load
           payload[:_source] = false
         end
-
-        if options[:type] || (klass != searchkick_klass && searchkick_index)
-          @type = [options[:type] || klass].flatten.map { |v| searchkick_index.klass_document_type(v) }
-        end
-
-        # routing
-        @routing = options[:routing] if options[:routing]
       end
+
+      # type
+      if options[:type] || (klass != searchkick_klass && searchkick_index)
+        @type = [options[:type] || klass].flatten.map { |v| searchkick_index.klass_document_type(v) }
+      end
+
+      # routing
+      @routing = options[:routing] if options[:routing]
 
       # merge more body options
       payload = payload.deep_merge(options[:body_options]) if options[:body_options]
@@ -489,7 +500,8 @@ module Searchkick
 
     def set_fields
       boost_fields = {}
-      fields = options[:fields] || searchkick_options[:searchable]
+      fields = options[:fields] || searchkick_options[:default_fields] || searchkick_options[:searchable]
+      all = searchkick_options.key?(:_all) ? searchkick_options[:_all] : below60?
       default_match = options[:match] || searchkick_options[:match] || :word
       fields =
         if fields
@@ -500,12 +512,12 @@ module Searchkick
             boost_fields[field] = boost.to_f if boost
             field
           end
-        elsif default_match == :word
+        elsif all && default_match == :word
           ["_all"]
-        elsif default_match == :phrase
+        elsif all && default_match == :phrase
           ["_all.phrase"]
         else
-          raise ArgumentError, "Must specify fields"
+          raise ArgumentError, "Must specify fields to search"
         end
       [boost_fields, fields]
     end
@@ -575,12 +587,18 @@ module Searchkick
       payload[:indices_boost] = indices_boost
     end
 
-    def set_suggestions(payload)
-      suggest_fields = (searchkick_options[:suggest] || []).map(&:to_s)
+    def set_suggestions(payload, suggest)
+      suggest_fields = nil
 
-      # intersection
-      if options[:fields]
-        suggest_fields &= options[:fields].map { |v| (v.is_a?(Hash) ? v.keys.first : v).to_s.split("^", 2).first }
+      if suggest.is_a?(Array)
+        suggest_fields = suggest
+      else
+        suggest_fields = (searchkick_options[:suggest] || []).map(&:to_s)
+
+        # intersection
+        if options[:fields]
+          suggest_fields &= options[:fields].map { |v| (v.is_a?(Hash) ? v.keys.first : v).to_s.split("^", 2).first }
+        end
       end
 
       if suggest_fields.any?
@@ -592,6 +610,8 @@ module Searchkick
             }
           }
         end
+      else
+        raise ArgumentError, "Must pass fields to suggest option"
       end
     end
 
@@ -822,7 +842,7 @@ module Searchkick
         if value.any?(&:nil?)
           {bool: {should: [term_filters(field, nil), term_filters(field, value.compact)]}}
         else
-          {in: {field => value}}
+          {terms: {field => value}}
         end
       elsif value.nil?
         {bool: {must_not: {exists: {field: field}}}}
@@ -863,13 +883,21 @@ module Searchkick
           }
         }
 
-        {
-          filter: {
+        if value[:missing]
+          if below50?
+            raise ArgumentError, "The missing option for boost_by is not supported in Elasticsearch < 5"
+          else
+            script_score[:field_value_factor][:missing] = value[:missing].to_f
+          end
+        else
+          script_score[:filter] = {
             exists: {
               field: field
             }
           }
-        }.merge(script_score)
+        end
+
+        script_score
       end
     end
 
@@ -896,6 +924,10 @@ module Searchkick
 
     def below50?
       Searchkick.server_below?("5.0.0-alpha1")
+    end
+
+    def below60?
+      Searchkick.server_below?("6.0.0-alpha1")
     end
   end
 end