+++ test/index_test.rb
@@ -59,6 +59,10 @@ class IndexTest < Minitest::Test
     assert_equal ["Dollar Tree"], Store.search(body: {query: {match: {name: "Dollar Tree"}}}, load: false).map(&:name)
   end
 
+  def test_body_warning
+    assert_output(nil, "The body option replaces the entire body, so the following options are ignored: where\n") { Store.search(body: {query: {match: {name: "dollar"}}}, where: {id: 1}) }
+  end
+
   def test_block
     store_names ["Dollar Tree"]
     products =
@@ -132,13 +136,22 @@ class IndexTest < Minitest::Test
     assert_search "*", [], where: {alt_description: "Hello"}
   end
 
+  def test_filterable_non_string
+    store [{name: "Product A", store_id: 1}]
+    assert_search "*", ["Product A"], where: {store_id: 1}
+  end
+
   def test_large_value
     skip if nobrainer?
     large_value = 1000.times.map { "hello" }.join(" ")
     store [{name: "Product A", text: large_value}], Region
     assert_search "product", ["Product A"], {}, Region
     assert_search "hello", ["Product A"], {fields: [:name, :text]}, Region
-    assert_search "hello", ["Product A"], {}, Region
+
+    # needs fields for ES 6
+    if elasticsearch_below60?
+      assert_search "hello", ["Product A"], {}, Region
+    end
   end
 
   def test_very_large_value