+++ test/sql_test.rb
@@ -195,4 +195,20 @@ class SqlTest < Minitest::Test
     store_names ["Product A"]
     assert Product.search("product", includes: [:store]).first.association(:store).loaded?
   end
+
+  def test_model_includes
+    skip unless defined?(ActiveRecord)
+
+    store_names ["Product A"]
+    store_names ["Store A"], Store
+
+    associations = {Product => [:store], Store => [:products]}
+    result = Searchkick.search("*", index_name: [Product, Store], model_includes: associations)
+
+    assert_equal 2, result.length
+
+    result.group_by(&:class).each_pair do |klass, records|
+      assert records.first.association(associations[klass].first).loaded?
+    end
+  end
 end