+++ test/test_helper.rb
@@ -42,6 +42,10 @@ def elasticsearch_below50?
   Searchkick.server_below?("5.0.0-alpha1")
 end
 
+def elasticsearch_below60?
+  Searchkick.server_below?("6.0.0-alpha1")
+end
+
 def elasticsearch_below22?
   Searchkick.server_below?("2.2.0")
 end
@@ -111,6 +115,12 @@ if defined?(Mongoid)
 
   class Cat < Animal
   end
+
+  class Sku
+    include Mongoid::Document
+
+    field :name
+  end
 elsif defined?(NoBrainer)
   NoBrainer.configure do |config|
     config.app_name = :searchkick
@@ -170,6 +180,13 @@ elsif defined?(NoBrainer)
   end
 
   class Cat < Animal
+  end
+
+  class Sku
+    include NoBrainer::Document
+
+    field :id,   type: String
+    field :name, type: String
   end
 elsif defined?(Cequel)
   cequel =
@@ -250,6 +267,13 @@ elsif defined?(Cequel)
   end
 
   class Cat < Animal
+  end
+
+  class Sku
+    include Cequel::Record
+
+    key :id, :uuid
+    column :name, :text
   end
 
   [Product, Store, Region, Speaker, Animal].each(&:synchronize_schema)
@@ -339,6 +363,10 @@ else
     t.string :type
   end
 
+  ActiveRecord::Migration.create_table :skus, id: :uuid do |t|
+    t.string :name
+  end
+
   class Product < ActiveRecord::Base
     belongs_to :store
   end
@@ -360,6 +388,9 @@ else
   end
 
   class Cat < Animal
+  end
+
+  class Sku < ActiveRecord::Base
   end
 end
 
@@ -417,6 +448,7 @@ end
 
 class Store
   searchkick \
+    default_fields: elasticsearch_below60? ? nil : [:name],
     routing: true,
     merge_mappings: true,
     mappings: {
@@ -438,6 +470,7 @@ end
 
 class Region
   searchkick \
+    default_fields: elasticsearch_below60? ? nil : [:name],
     geo_shape: {
       territory: {tree: "quadtree", precision: "10km"}
     }
@@ -455,6 +488,7 @@ end
 
 class Speaker
   searchkick \
+    default_fields: elasticsearch_below60? ? nil : [:name],
     conversions: ["conversions_a", "conversions_b"]
 
   attr_accessor :conversions_a, :conversions_b, :aisle
@@ -470,11 +504,16 @@ end
 
 class Animal
   searchkick \
+    default_fields: elasticsearch_below60? ? nil : [:name],
     text_start: [:name],
     suggest: [:name],
     index_name: -> { "#{name.tableize}-#{Date.today.year}#{Searchkick.index_suffix}" },
     callbacks: defined?(ActiveJob) ? :async : true
     # wordnet: true
+end
+
+class Sku
+  searchkick callbacks: defined?(ActiveJob) ? :async : true
 end
 
 Product.searchkick_index.delete if Product.searchkick_index.exists?
@@ -493,6 +532,7 @@ class Minitest::Test
     Store.destroy_all
     Animal.destroy_all
     Speaker.destroy_all
+    Sku.destroy_all
   end
 
   protected